Collisions bugfix
Fixed a bug with collisions where the player could move a piece over the edge of the board with the snapping mechanic. Code now checks for collisions in the update method, and only moves the object if it isn't being dragged, e.g. if on snapping it entered an illegal position.